refactor(dto): move BenefitItem, MenuItem into dto package
