--- src/main/java/christmas/domain/MenuItem.java
@@ -1,4 +1,0 @@
-package christmas.domain;
-
-public record MenuItem(Menu menu, int quantity) {
-}
