+++ src/main/java/christmas/dto/BenefitItem.java
@@ -1,4 +1,4 @@
-package christmas.domain;
+package christmas.dto;
 
 public record BenefitItem(String name, int amount) {
 }
