+++ src/main/java/christmas/dto/MenuItem.java
@@ -1,0 +1,6 @@
+package christmas.dto;
+
+import christmas.domain.Menu;
+
+public record MenuItem(Menu menu, int quantity) {
+}
